add viewcount

+++ src/main/resources/db/migration/V1__create_user_table.sql
@@ -1,6 +1,6 @@
   create table USER
         (
-          ID INT auto_increment  primary key,
+          ID bigint auto_increment  primary key,
           ACCOUNT_ID   VARCHAR(100),
           NAME         VARCHAR(50),
           TOKEN        CHAR(36),
